requirements.txt
pandas
psycopg2-binary
geopy
requests
